requirements.txt
# ==========================================
# EVALUAI PROFESOR - DEPENDENCIAS
# ==========================================

# Framework base (AgentSquad)
agent-squad[openai]>=0.1.0

# API y Servidor
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# DeepSeek / OpenAI API
openai>=1.12.0

# Base de datos
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Procesamiento de documentos
PyPDF2>=3.0.1
python-docx>=1.1.0

# Utilidades
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
httpx>=0.26.0
